requirements.txt
fastapi 
uvicorn 
pypdf 
pdfplumber
sentence-transformers 
faiss-cpu 
langchain 
langchain-community 
python-dotenv